azure/jenkins-controller: switch to casc
The configuration-as-code plugin is the recommended way to configure
Jenkins, not the Groovy script. Most plugin settings seem to be
available through that.

The yaml was produced by checking in the dump created by the plugin.

As per jenkinsci/configuration-as-code-plugin#393, it is not possible to
disable the setup wizard from casc itself.

It is however also not possible to use the option suggested there to
disable it as we *need* the `initialAdminPassword` to install
plugins (including casc) in first place, and if this is set,
`initialAdminPassword` won't ever be created.

Break the circle by starting Jenkins with initial setup enabled, use
it to install all plugins, then use the groovy script to disable the
initial setup. Afterwards, restart Jenkins to deploy all changes and
plugins.

Signed-off-by: Florian Klink <flokli@flokli.de>